Update throughput test to work with current dgraph
The throughput test in its current state seems to only work for older
versions of dgraph which supported xid. It also looks like the expected
return values from the server have changed a bit since this benchmark
was written. I have updated the queries and the expected result fields
to work with the current version of dgraph.

Because dgraph no longer accepts XIDs, one has to use UIDs. I have
included a few scripts to scrape the appropriate UIDs from a populated
dgraph database and dump them into files with the format the benchmark
expects.

Also added a small readme on usage.