Set Grafana dashboard data source as an input source.
Previously the dashboard assumed that the Prometheus data source with all the
Dgraph metrics was named "Dgraph". This made it difficult to import if the
Prometheus data source was named something else.

Now when the template is imported it will ask the admin to choose the data
source that has the Dgraph metrics instead of assuming the data source is named
"Dgraph".