Sync Schemas and add pred "email"
Just by synchronizing schemas and adding the "email" predicate to the schemas.

The reason for the pred "email" is in case a Play Dgraph user comes to test "upsert" on it. It does not receive errors like ": Attribute email is not indexed."

At RDF 21milion you have 2 emails sfdowntown@hiusa.org and reservations@hotelgriffon.com. That is, this change does not change anything significant. And it could be used in place of "bob@example.com" in docs.